+++ R/response_rate.R
@@ -14,11 +14,12 @@ num_respondents <- function(df) { NROW(df) }
 #' @param na. NA responses in the dataframe are coded as this.
 #' @export
 num_answers <- function(df, variable, na = NA) {
-  num_answers_(df, deparse(substitute(variable)), na = na)
+  if (!is.character(substitute(variable))) { variable <- deparse(substitute(variable)) }
+  num_answers_(df, variable, na = na)
 }
 
 num_answers_ <- function(df, variable, na = NA) {
-  df %>% gather_(variable) %>% na.rm %>% num_respondents
+  df %>% select_(variable) %>% na.rm %>% num_respondents
 }
 
 
@@ -29,4 +30,7 @@ num_answers_ <- function(df, variable, na = NA) {
 #' @param variable character. The variable to count.
 #' @param na. NA responses in the dataframe are coded as this.
 #' @export
-response_rate <- function(df, variable, na = NA) num_answers(df, variable, na) / num_respondents(df)
+response_rate <- function(df, variable, na = NA) {
+  if (!is.character(substitute(variable))) { variable <- deparse(substitute(variable)) }
+  num_answers_(df, variable, na) / num_respondents(df)
+}